moving registry creds test back to addons test, adding an extra top level 'addons' in config file
