+++ test/integration/addons_test.go
@@ -100,7 +100,7 @@ func TestAddons(t *testing.T) {
 		// so we override that here to let minikube auto-detect appropriate cgroup driver
 		os.Setenv(constants.MinikubeForceSystemdEnv, "")
 
-		args := append([]string{"start", "-p", profile, "--wait=true", "--memory=4000", "--alsologtostderr", "--addons=registry", "--addons=metrics-server", "--addons=volumesnapshots", "--addons=csi-hostpath-driver", "--addons=gcp-auth", "--addons=cloud-spanner", "--addons=inspektor-gadget", "--addons=nvidia-device-plugin", "--addons=yakd", "--addons=volcano", "--addons=amd-gpu-device-plugin"}, StartArgs()...)
+		args := append([]string{"start", "-p", profile, "--wait=true", "--memory=4000", "--alsologtostderr", "--addons=registry", "--addons=registry-creds", "--addons=metrics-server", "--addons=volumesnapshots", "--addons=csi-hostpath-driver", "--addons=gcp-auth", "--addons=cloud-spanner", "--addons=inspektor-gadget", "--addons=nvidia-device-plugin", "--addons=yakd", "--addons=volcano", "--addons=amd-gpu-device-plugin"}, StartArgs()...)
 		if !NoneDriver() {
 			args = append(args, "--addons=ingress", "--addons=ingress-dns", "--addons=storage-provisioner-rancher")
 		}
@@ -140,6 +140,7 @@ func TestAddons(t *testing.T) {
 	t.Run("parallel", func(t *testing.T) {
 		tests := []TestCase{
 			{"Registry", validateRegistryAddon},
+			{"RegistryCreds", validateRegistryCredsAddon},
 			{"Ingress", validateIngressAddon},
 			{"InspektorGadget", validateInspektorGadgetAddon},
 			{"MetricsServer", validateMetricsServerAddon},
@@ -301,6 +302,64 @@ func validateIngressAddon(ctx context.Context, t *testing.T, profile string) {
 	// nslookup should include info about the hello-john.test host, including minikube's ip
 	if !strings.Contains(rr.Stdout.String(), ip) {
 		t.Errorf("unexpected output from nslookup. stdout: %v\nstderr: %v", rr.Stdout.String(), rr.Stderr.String())
+	}
+}
+
+// validateRegistryCredsAddon tests the registry-creds addon by trying to load its configs
+func validateRegistryCredsAddon(ctx context.Context, t *testing.T, profile string) {
+	defer disableAddon(t, "registry-creds", profile)
+	defer PostMortemLogs(t, profile)
+
+	client, err := kapi.Client(profile)
+	if err != nil {
+		t.Fatalf("failed to get Kubernetes client for %s : %v", profile, err)
+	}
+
+	start := time.Now()
+	if err := kapi.WaitForDeploymentToStabilize(client, "kube-system", "registry-creds", Minutes(6)); err != nil {
+		t.Errorf("failed waiting for registry-creds deployment to stabilize: %v", err)
+	}
+	t.Logf("registry-creds stabilized in %s", time.Since(start))
+
+	rr, err := Run(t, exec.CommandContext(ctx, Target(), "addons", "configure", "registry-creds", "-f", "./testdata/addons_testconfig.json", "-p", profile))
+	if err != nil {
+		t.Errorf("failed to configure addon. args %q : %v", rr.Command(), err)
+	}
+
+	// Check a few secrets exists that match our test data
+	// In our test aws and gcp are set, docker and acr are disabled - so they will be set to "changeme"
+	rr, err = Run(t, exec.CommandContext(ctx, "kubectl", "--context", profile, "-n", "kube-system", "get", "secret", "-o", "yaml"))
+	if err != nil {
+		t.Errorf("failed to get secrets. args %q : %v", rr.Command(), err)
+	}
+
+	expected := []string{
+		"DOCKER_PRIVATE_REGISTRY_PASSWORD: Y2hhbmdlbWU=",
+		"DOCKER_PRIVATE_REGISTRY_SERVER: Y2hhbmdlbWU=",
+		"DOCKER_PRIVATE_REGISTRY_USER: Y2hhbmdlbWU=",
+
+		"ACR_CLIENT_ID: Y2hhbmdlbWU=",
+		"ACR_PASSWORD: Y2hhbmdlbWU=",
+		"ACR_URL: Y2hhbmdlbWU=",
+
+		"AWS_ACCESS_KEY_ID: dGVzdF9hd3NfYWNjZXNzaWQ=",
+		"AWS_SECRET_ACCESS_KEY: dGVzdF9hd3NfYWNjZXNza2V5",
+		"AWS_SESSION_TOKEN: dGVzdF9hd3Nfc2Vzc2lvbl90b2tlbg==",
+		"aws-account: dGVzdF9hd3NfYWNjb3VudA==",
+		"aws-assume-role: dGVzdF9hd3Nfcm9sZQ==",
+		"aws-region: dGVzdF9hd3NfcmVnaW9u",
+
+		"application_default_credentials.json: ewogICJjbGllbnRfaWQiOiAiaGFoYSIsCiAgImNsaWVudF9zZWNyZXQiOiAibmljZV90cnkiLAogICJxdW90YV9wcm9qZWN0X2lkIjogInRoaXNfaXNfZmFrZSIsCiAgInJlZnJlc2hfdG9rZW4iOiAibWF5YmVfbmV4dF90aW1lIiwKICAidHlwZSI6ICJhdXRob3JpemVkX3VzZXIiCn0K",
+		"gcrurl: aHR0cHM6Ly9nY3IuaW8=",
+	}
+
+	rrout := strings.TrimSpace(rr.Stdout.String())
+	for _, exp := range expected {
+		re := regexp.MustCompile(fmt.Sprintf(".*%s.*", exp))
+		secret := re.FindString(rrout)
+		if secret == "" {
+			t.Errorf("Did not find expected secret: '%s'", secret)
+		}
 	}
 }
